+++ packages/desktop/src/rtl-tests/community.join.test.tsx
@@ -20,7 +20,6 @@ import {
   SocketActionTypes,
   RegisterUserCertificatePayload,
   InitCommunityPayload,
-  Community,
   ErrorCodes,
   ErrorMessages,
   getFactory,
@@ -32,18 +31,29 @@ import LoadingPanel from '../renderer/components/LoadingPanel/LoadingPanel'
 import { createUserCertificateTestHelper } from '@quiet/identity'
 import { AnyAction } from 'redux'
 import {
-  ChannelsReplicatedPayload,
-  ErrorPayload,
-  ResponseLaunchCommunityPayload,
-  SendOwnerCertificatePayload,
+  type InvitationData,
+  type ChannelsReplicatedPayload,
+  type Community,
+  type ErrorPayload,
+  type ResponseLaunchCommunityPayload,
+  type SendOwnerCertificatePayload,
 } from '@quiet/types'
+import { composeInvitationShareUrl } from '@quiet/common'
 
 jest.setTimeout(20_000)
 
 describe('User', () => {
   let socket: MockedSocket
-  const validCode =
-    'QmZoiJNAvCffeEHBjk766nLuKVdkxkAT7wfFJDPPLsbKSE=y7yczmugl2tekami7sbdz5pfaemvx7bahwthrdvcbzw5vex2crsr26qd'
+  const validData: InvitationData = {
+    pairs: [
+      {
+        onionAddress: 'y7yczmugl2tekami7sbdz5pfaemvx7bahwthrdvcbzw5vex2crsr26qd',
+        peerId: 'QmZoiJNAvCffeEHBjk766nLuKVdkxkAT7wfFJDPPLsbKSE',
+      },
+    ],
+    psk: 'BNlxfE2WBF7LrlpIX0CvECN5o1oZtA16PkAb7GYiwYw=',
+  }
+  const validCode = composeInvitationShareUrl(validData)
   // trigger
   beforeEach(() => {
     socket = new MockedSocket()
@@ -195,6 +205,7 @@ describe('User', () => {
       Array [
         "Communities/createNetwork",
         "Communities/setInvitationCodes",
+        "Communities/savePSK",
         "Communities/addNewCommunity",
         "Communities/setCurrentCommunity",
         "Modals/closeModal",