+++ packages/mobile/src/store/init/deepLink/deepLink.saga.test.ts
@@ -3,20 +3,20 @@ import { combineReducers } from '@reduxjs/toolkit'
 import { reducers } from '../../root.reducer'
 import { Store } from '../../store.types'
 import { prepareStore } from '../../../tests/utils/prepareStore'
-import { communities, Community, connection, getInvitationCodes, identity } from '@quiet/state-manager'
+import { communities, connection, identity } from '@quiet/state-manager'
 import { initActions } from '../init.slice'
 import { navigationActions } from '../../navigation/navigation.slice'
 import { ScreenNames } from '../../../const/ScreenNames.enum'
 import { deepLinkSaga } from './deepLink.saga'
-import { CommunityOwnership, ConnectionProcessInfo, Identity } from '@quiet/types'
+import { type Community, CommunityOwnership, ConnectionProcessInfo, type Identity, InvitationData } from '@quiet/types'
+import { composeInvitationShareUrl, validInvitationUrlTestData } from '@quiet/common'
 
 describe('deepLinkSaga', () => {
   let store: Store
 
   const id = '00d045ab'
-  const validCode =
-    'QmZoiJNAvCffeEHBjk766nLuKVdkxkAT7wfFJDPPLsbKSE=y7yczmugl2tekami7sbdz5pfaemvx7bahwthrdvcbzw5vex2crsr26qd'
-  const validPairs = getInvitationCodes(validCode)
+  const validData: InvitationData = validInvitationUrlTestData.data
+  const validCode = validInvitationUrlTestData.code()
   const community: Community = {
     id,
     name: '',
@@ -70,7 +70,8 @@ describe('deepLinkSaga', () => {
       .put(
         communities.actions.createNetwork({
           ownership: CommunityOwnership.User,
-          peers: validPairs,
+          peers: validData.pairs,
+          psk: validData.psk,
         })
       )
       .run()
@@ -104,7 +105,8 @@ describe('deepLinkSaga', () => {
       .not.put(
         communities.actions.createNetwork({
           ownership: CommunityOwnership.User,
-          peers: validPairs,
+          peers: validData.pairs,
+          psk: validData.psk,
         })
       )
       .run()
@@ -124,10 +126,65 @@ describe('deepLinkSaga', () => {
     await expectSaga(deepLinkSaga, initActions.deepLink(validCode))
       .withReducer(reducer)
       .withState(store.getState())
+      .put.like({
+        action: {
+          type: navigationActions.replaceScreen.type,
+          payload: {
+            screen: ScreenNames.ErrorScreen,
+            params: {
+              title: 'You already belong to a community',
+              message: "We're sorry but for now you can only be a member of a single community at a time",
+            },
+          },
+        },
+      })
       .not.put(
         communities.actions.createNetwork({
           ownership: CommunityOwnership.User,
-          peers: validPairs,
+          peers: validData.pairs,
+          psk: validData.psk,
+        })
+      )
+      .run()
+  })
+
+  test('displays error if invitation code is invalid', async () => {
+    const invalidData: InvitationData = {
+      pairs: [
+        {
+          onionAddress: 'y7yczmugl2tekami7sbdz5pfaemvx7bahwthrdvcbzw5vex2crsr26qd',
+          peerId: 'QmZoiJNAvCffeEHBjk766nLuKVdkxkAT7wfFJDPPLsbKSE',
+        },
+      ],
+      psk: 'BNlxfE=',
+    }
+    const invalidCode = composeInvitationShareUrl(invalidData)
+    store.dispatch(
+      initActions.setWebsocketConnected({
+        dataPort: 5001,
+      })
+    )
+    const reducer = combineReducers(reducers)
+    await expectSaga(deepLinkSaga, initActions.deepLink(invalidCode))
+      .withReducer(reducer)
+      .withState(store.getState())
+      .put.like({
+        action: {
+          type: navigationActions.replaceScreen.type,
+          payload: {
+            screen: ScreenNames.ErrorScreen,
+            params: {
+              title: 'Invalid invitation link',
+              message: 'Please check your invitation link and try again',
+            },
+          },
+        },
+      })
+      .not.put(
+        communities.actions.createNetwork({
+          ownership: CommunityOwnership.User,
+          peers: validData.pairs,
+          psk: validData.psk,
         })
       )
       .run()
@@ -166,7 +223,8 @@ describe('deepLinkSaga', () => {
       .not.put(
         communities.actions.createNetwork({
           ownership: CommunityOwnership.User,
-          peers: validPairs,
+          peers: validData.pairs,
+          psk: validData.psk,
         })
       )
       .run()